feat: `Square` type aliases moved to `colour_and_aliases`
